+++ shraga_common/app/__init__.py
@@ -1,6 +1,5 @@
 from .config import get_config
 from .exceptions import RequestCancelledException
 from .routers import setup_app
-from .services import list_flows_service
 
-__all__ = ['RequestCancelledException', 'setup_app', 'get_config', 'list_flows_service']
+__all__ = ['RequestCancelledException', 'setup_app', 'get_config']
